requirements.txt
azure-core==1.32.0
azure-functions==1.21.3
azure-identity==1.19.0
azure-keyvault==4.2.0
azure-keyvault-certificates==4.9.0
azure-keyvault-keys==4.10.0
azure-keyvault-secrets==4.9.0
azure-storage-blob==12.24.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
et_xmlfile==2.0.0
idna==3.10
isodate==0.7.2
msal==1.31.1
msal-extensions==1.2.0
numpy==2.2.2
openpyxl==3.1.5
pandas==2.2.3
portalocker==2.10.1
pycparser==2.22
PyJWT==2.10.1
pyodbc==5.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
pyxlsb==1.0.10
requests==2.32.3
six==1.17.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
xlrd==2.0.1
